i3mac: gigantic refactor of entire codebase
- Introduce the notion of physical and virtual displays, to mimic displays
  on i3. Reworked moving windows between displays/focusing a display to
  copy i3 behaviour. This includes moving windows to virtual displays that
  currently don't exist.
- Status bars are now managed by a physical display, rather than the
  window manager itself, and it is a physical display's responsibility to
  update its status bar.
- Big reduction in the amount of code written for the WindowManger, Display
  (now VirtualDisplay) and Container, by being more deliberate about what a
  certain action should perform, and doing it smartly, rather than, for
  example, huge copy-pasted blocks of new Bounds creation.
- Fixed some issues with window resizing.

Still requires some clean-up effort, especially still in the Container
struct but functionality is getting there.

+++ core_graphics/src/keyboard.rs
@@ -153,7 +153,7 @@ impl TryFrom<i64> for Keycode {
     }
 }
 
-#[derive(Debug)]
+#[derive(Copy, Clone, Debug)]
 pub enum Direction {
     Left,
     Right,
@@ -166,7 +166,7 @@ pub enum KeyCommand {
     NewTerminal,
     CloseWindow,
     Focus(Direction),
-    FocusDisplay(u64),
+    FocusDisplay(usize),
     MoveWindowToDisplay(u64),
     ToggleVerticalSplit,
     ToggleHorizontalSplit,
